+++ PickFLPrj/src/com/pickfl/main/controller/MainPageController.java
@@ -1,4 +1,4 @@
-package com.kh.main.controller;
+package com.pickfl.main.controller;
 
 import java.io.IOException;
